Fix python modules hierarchy
