+++ python/tweedledum/classical/kitty.cpp
