+++ python/tweedledum/ir/__init__.py
@@ -1,0 +1,5 @@
+#-------------------------------------------------------------------------------
+# Part of Tweedledum Project.  This file is distributed under the MIT License.
+# See accompanying file /LICENSE for details.
+#-------------------------------------------------------------------------------
+from .._tweedledum.ir import *
