+++ python/tweedledum/operators/extension.cpp
@@ -2,7 +2,7 @@
 | Part of Tweedledum Project.  This file is distributed under the MIT License.
 | See accompanying file /LICENSE for details.
 *-----------------------------------------------------------------------------*/
-#include "Operators.h"
+#include "operators.h"
 
 #include <kitty/kitty.hpp>
 #include <mockturtle/networks/xag.hpp>
