+++ python/tweedledum/operators/operators.h
